Login model moved to auth and renamed to LoginResponse
